Unzipping plugins error fixed.

+++ wordpress-installer.sh
@@ -1,22 +1,5 @@
 #!/bin/bash
-#set -xv
-
-#Sudo girişi yapılıyor
-#if [[ ! $(id -u) -eq 0 ]]; then
-#	while [ 1 ]; do
-#		s=$(zenity --password | sudo -S echo "HOŞGELDİNİZ!");
-#		if [ $? == "0" ]; then
-#			break;
-#		fi
-#	done
-#fi 
-
-
-#Kurulu değil ise LFTP kur
-if [ ! "command -v lftp" ]; then
-	$sistem "sudo apt-get -y install lftp";
-fi
-
+set -xv
 notify-send "Welcome to Wordpress Installation!"
 
 #FORM İLE BİLGİLER ALINIYOR
@@ -30,6 +13,11 @@ setup_inf=$(zenity \
 	--add-entry="Ftp Server" \
 	--add-entry="Ftp Username" \
 	--add-entry="Ftp User Password")
+#Check Cancel
+if [ $setup_inf = 1 ]; then
+	exit;
+fi
+
 #VERİTABANI BİLGİLERİ
 DBNAME=$(echo $setup_inf | cut -d':' -f1);
 DBUSER=$(echo $setup_inf | cut -d':' -f2);
@@ -43,19 +31,19 @@ DOMAIN=$(echo $setup_inf | cut -d':' -f5 | cut -d'@' -f1);
 
 #YÜKLENECEK EKLENTİLERİN SEÇİLMESİ
 wp_eklentileri=$(zenity  \
-	--width=500 --height=400 --list --text "Wordpress'e eklemek istediğiniz eklentileri seçin!" \
-	--checklist --column "Seç" --column "Eklentiler" --column "Açıklama" --separator=":" \
-	FALSE "All in one Seo Pack" "Seo Eklentisi"\
-	FALSE "Google Xml Sitemap" "Otomatik Sitemap Oluşturma"\
-	FALSE "Wp Optimize" "Wordpress Optimize Eklentisi"\
-	FALSE "AG Custom Admin" "Admin Girişi Özelleştirme"\
-	FALSE "Ozh Admin Drop Down Menu" "Admin Paneli Özelleştirme"\
-	FALSE "Codestyling localization" "Sitedeki Dil Özelleştirme"\
-	FALSE "Maintenance Mode" "Bakım Modu"\
-	FALSE "Wp Google Fonts" "Font Özelleştirme"\
-	FALSE "Page Links To" "Sayfalara Özel Link Ekleme"\
-	FALSE "Nextgen Galeri" "Galeri Eklentisi"\
-	FALSE "Category and Page Icons" "Kategori ve Sayfa ikonları"
+	--width=1100 --height=400 --list --text "Choose the plugins for adding to wordpress!" \
+	--checklist --column "Choose" --column "Plugins" --column "Description" --separator=":" \
+	FALSE "All in one Seo Pack" "All in One SEO Pack is a WordPress SEO plugin to automatically optimize your WordPress blog for Search Engines such as Google."\
+	FALSE "Google Xml Sitemap" "In its simplest terms, a XML Sitemap-usually called Sitemap, with a capital S-is a list of the pages on your website."\
+	FALSE "Wp Optimize" "Simple but effective plugin allows you to extensively clean up your WordPress database and optimize it without doing manual queries."\
+	FALSE "AG Custom Admin" "All-in-one tool for admin panel customization. Change almost everything: admin menu, dashboard, login page, admin bar etc."\
+	FALSE "Ozh Admin Drop Down Menu" "All admin links available in a neat horizontal drop down menu. Saves lots of screen real estate!"\
+	FALSE "Codestyling localization" "You can manage and edit all gettext translation files (*.po/*.mo) in WordPress Admin Center."\
+	FALSE "Maintenance Mode" "Adds a splash page to your site that lets visitors know your site is down for maintenance. "\
+	FALSE "Wp Google Fonts" "The WP Google Fonts plugin allows you to easily add fonts from the Google Font Directory to your WordPress theme."\
+	FALSE "Page Links To" "Lets you make a WordPress page (or other content type) link to an external URL of your choosing, instead of its WordPress URL."\
+	FALSE "Nextgen Gallery" "The most popular WordPress gallery plugin and one of the most popular plugins of all time with over 10 million downloads."\
+	FALSE "Category and Page Icons" "Easy add icons to sidebar of categories and pages."
 	)
 e=1;
 for (( i = 1; i < 12; i++ )); do
@@ -84,38 +72,43 @@ for (( i = 1; i < 12; i++ )); do
 		eklenti[$e]="category-page-icons"; e=$((e+1));
 	fi
 done
+#Check Cancel
+if [ $wp_eklentileri = 1 ]; then
+	exit;
+fi
 
 
-#dosya yoksa indir
-notify-send "Kurulum işlemi başlıyor!"
-#Türkçe son versiyon indiriliyor
-VERSIYON=$(lynx --source http://tr.wordpress.org/releases/#latest | grep -m 1 "<td>" | cut -d'>' -f3 | cut -d'<' -f1);
-wget http://tr.wordpress.org/wordpress-$VERSIYON-tr_TR.tar.gz;
-#Arşivden çıkartılıyor
-tar -zxvf wordpress-$VERSIYON-tr_TR.tar.gz;
+notify-send "Downloading files!"
+#Dosya yoksa indir
+if [ ! -f ./latest.zip ]; then
+wget https://wordpress.org/latest.zip;
+fi
+#Unzip the file
+unzip latest.zip;
 #Wordpress klasörüne geçiliyor
 cd wordpress
 #Veritabanı bilgilieri dosyaya ekleniyor
-sed "s/veritabaniismi/$DBNAME/g" wp-config-sample.php > wp-config-sample-1.php;
-sed "s/kullaniciadi/$DBUSER/g" wp-config-sample-1.php > wp-config-sample-2.php;
-sed "s/parola/$DBPASS/g" wp-config-sample-2.php > wp-config.php;
+sed "s/database_name_here/$DBNAME/g" wp-config-sample.php > wp-config-sample-1.php;
+sed "s/username_here/$DBUSER/g" wp-config-sample-1.php > wp-config-sample-2.php;
+sed "s/password_here/$DBPASS/g" wp-config-sample-2.php > wp-config.php;
 rm wp-config-sample-1.php wp-config-sample-2.php
 
 
 #Eklentiler klasöre ekleniyor
 cd wp-content/plugins &&
 for (( i = 1; i < $e; i++ )); do
+	echo "i="$i;
+	echo "e="$e;
 	wget http://downloads.wordpress.org/plugin/${eklenti[$i]}.zip
 done
 if [ "ls | *.zip" ]; then
-	unzip *.zip;
+	unzip "*.zip";
 	rm *.zip;
 fi
 
-
 #Bulunduğum dizinde ne var ne yoksa yükle
-notify-send "Ftp'ye yükleme işlemi başladı!"
-SOURCEFOLDER="../wordpress"
+notify-send "Files are uploading to server!"
+SOURCEFOLDER="../../wordpress"
 TARGETFOLDER="/"
 lftp -f "
 open $FTPHOST
@@ -127,9 +120,9 @@ bye
 
 
 #Gereksiz dosya klasörler siliniyor
-rm -rf ../../../wordpres*
-notify-send "Sitenin kurulumu tamamlandı!"
-zenity --text="Admin Paneli Açılsın mı?" --question;
-if [ "$?" == "0" ]; then google-chrome http://www.$DOMAIN.com/wp-admin; fi
+#rm -rf ../../../wordpress
+notify-send "Wordpress installation finished!"
+zenity --text="Do you want to open admin panel?" --question;
+if [ "$?" == "0" ]; then sensible-browser http://www.$DOMAIN.com/wp-admin & ; fi
 exit;
